\name{lipdbf}
\alias{lipdbf}

\docType{data}

\title{
Shapefile (.dbf) for the 56 districts in Scotland for the lip cancer data
}

\description{
lipdbf contains the .dbf component of the shapefile for the 56 districts in 
Scotland for the lip cancer data. These data  are used in a  worked example in 
the vignette accompanying the CARBayes package.
}

\usage{data(lipdbf)}


\examples{
library(shapefiles)
data(lipdbf)
}


\keyword{datasets}
