\name{lipshp}
\alias{lipshp}

\docType{data}

\title{
Shapefile (.shp) for the 56 districts in Scotland for the lip cancer data
}

\description{
lipshp contains the .shp component of the shapefile for the 56 districts in 
Scotland for the lip cancer data. These data  are used in a  worked example in 
the vignette accompanying the CARBayes package.
}

\usage{data(lipshp)}



\examples{
library(shapefiles)
data(lipshp)
}


\keyword{datasets}
